# docs/swagger/swagger.yaml
basePath: /gateway
definitions:
...
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
...
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.NotificationRequest:
    properties:
      message:
        type: string
      target:
        type: string
      type:
        type: string
    type: object
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.NotificationResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      message:
        type: string
      target:
        type: string
      type:
        type: string
    type: object
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.PaymentRequest:
    properties:
      amount:
        type: number
      booking_id:
        type: string
      currency:
        type: string
    type: object
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.PaymentResponse:
    properties:
      id:
        type: string
      payment_url:
        type: string
      provider:
        type: string
      status:
        type: string
    type: object
...
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.RefundRequest:
    properties:
      amount:
        type: number
      payment_id:
        type: string
      reason:
        type: string
    type: object
  github_com_ftryyln_hotel-booking-microservices_pkg_dto.RefundResponse:
    properties:
      id:
        type: string
      reference:
        type: string
      status:
        type: string
    type: object
...
  paymenthttp.webhookResponse:
    properties:
      message:
        type: string
      payment_id:
        type: string
      status:
        type: string
    type: object
host: localhost:8088
info:
  contact: {}
  description: Aggregated API surface for hotel booking platform.
  title: Hotel Booking Microservices API
  version: "1.0"
paths:
...
  /notifications:
    get:
      parameters:
      - description: pagination limit (default 50)
        in: query
        name: limit
        type: integer
      - description: pagination offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.NotificationResponse'
            type: array
      summary: List notifications (in-memory)
      tags:
      - Notifications
    post:
      consumes:
      - application/json
      parameters:
      - description: Notification payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.NotificationRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.NotificationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
      summary: Send notification
      tags:
      - Notifications
  /notifications/{id}:
    get:
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.NotificationResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
      summary: Get notification by ID
      tags:
      - Notifications
  /payments:
    post:
      consumes:
      - application/json
      parameters:
      - description: Payment payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.PaymentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.PaymentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Initiate payment
      tags:
      - Payments
  /payments/{id}:
    get:
      parameters:
      - description: Payment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.PaymentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get payment by ID
      tags:
      - Payments
  /payments/by-booking/{booking_id}:
    get:
      parameters:
      - description: Booking ID
        in: path
        name: booking_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.PaymentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get payment by booking ID
      tags:
      - Payments
  /payments/refund:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refund payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.RefundRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.RefundResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Refund payment
      tags:
      - Payments
  /payments/webhook:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/paymenthttp.webhookResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ftryyln_hotel-booking-microservices_pkg_dto.ErrorResponse'
      summary: Payment webhook
      tags:
      - Payments
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"